PLANET-7611: Rename Happy Point
- File names updated
