+++ assets/src/frontendIndex.js
@@ -2,7 +2,7 @@ import 'regenerator-runtime/runtime';
 
 import {createRoot} from 'react-dom/client';
 import {TableOfContentsFrontend} from './blocks/TableOfContents/TableOfContentsFrontend';
-import {HappypointFrontend} from './blocks/Happypoint/HappypointFrontend';
+import {HappyPointFrontend} from './blocks/HappyPoint/HappyPointFrontend';
 import {MediaFrontend} from './blocks/Media/MediaFrontend';
 import {ColumnsFrontend} from './blocks/Columns/ColumnsFrontend';
 import {setupMediaElementJS} from './blocks/Media/setupMediaElementJS';
@@ -10,7 +10,7 @@ import {setupMediaElementJS} from './blocks/Media/setupMediaElementJS';
 // Render React components
 const COMPONENTS = {
   'planet4-blocks/submenu': TableOfContentsFrontend,
-  'planet4-blocks/happypoint': HappypointFrontend,
+  'planet4-blocks/happypoint': HappyPointFrontend,
   'planet4-blocks/media-video': MediaFrontend,
   'planet4-blocks/columns': ColumnsFrontend,
 };
