+++ assets/src/scss/blocks.scss
@@ -9,7 +9,7 @@
 @import "blocks/Cookies/CookiesStyle";
 @import "blocks/Counter/CounterStyle";
 @import "blocks/ENForm/ENFormStyle";
-@import "blocks/Happypoint/HappypointStyle";
+@import "blocks/HappyPoint/HappypointStyle";
 @import "blocks/Media";
 @import "blocks/SocialMedia/SocialMediaStyle";
 @import "blocks/Spreadsheet";
