+++ assets/src/scss/blocks/HappyPoint/HappypointEditorStyle.scss
