+++ assets/src/scss/blocks/HappyPoint/HappypointStyle.scss
