+++ src/Loader.php
@@ -205,7 +205,7 @@ final class Loader
         new Blocks\ENForm();//NOSONAR
         new Blocks\Gallery();//NOSONAR
         new Blocks\GuestBook();//NOSONAR
-        new Blocks\Happypoint();//NOSONAR
+        new Blocks\HappyPoint();//NOSONAR
         new Blocks\Media();//NOSONAR
         new Blocks\SocialMedia();//NOSONAR
         new Blocks\Spreadsheet();//NOSONAR
